+++ examples/with-eth-passkeys-galore/src/pages/index.tsx
@@ -2,8 +2,15 @@ import Image from "next/image";
 import { useForm } from "react-hook-form";
 import axios from "axios";
 import { useState, useEffect } from "react";
-import { Account, createWalletClient, http } from "viem";
+import { Account, createWalletClient, http, type WalletClient } from "viem";
+import { ethers } from "ethers";
 import { sepolia } from "viem/chains";
+import {
+  createSmartAccountClient,
+  BiconomySmartAccountV2,
+  PaymasterMode,
+  LightSigner,
+} from "@biconomy/account";
 
 import { TurnkeySigner } from "@turnkey/ethers";
 import { createAccount } from "@turnkey/viem";
@@ -16,8 +23,13 @@ type subOrgFormData = {
   subOrgName: string;
 };
 
-type signingFormData = {
+type signMessageFormData = {
   messageToSign: string;
+};
+
+type signTransactionFormData = {
+  destinationAddress: string;
+  amount: string;
 };
 
 type TWalletState = TWalletDetails | null;
@@ -36,14 +48,28 @@ export default function Home() {
 
   // Wallet is used as a proxy for logged-in state
   const [wallet, setWallet] = useState<TWalletState>(null);
+  const [smartAccountAddress, setSmartAccountAddress] = useState<string>("");
   const [signedMessage, setSignedMessage] = useState<TSignedMessage>(null);
+  const [signedTransaction, setSignedTransaction] = useState<string | null>(
+    null
+  );
   const [useViem, setUseViem] = useState(true);
+  const [viemClient, setViemClient] = useState<WalletClient | null>(null);
+  const [ethersClient, setEthersClient] = useState<TurnkeySigner | null>(null);
 
   const { handleSubmit: subOrgFormSubmit } = useForm<subOrgFormData>();
   const { register: signingFormRegister, handleSubmit: signingFormSubmit } =
-    useForm<signingFormData>();
+    useForm<signMessageFormData>();
   const { register: _loginFormRegister, handleSubmit: loginFormSubmit } =
     useForm();
+  const {
+    register: signTransactionFormRegister,
+    handleSubmit: signTransactionFormSubmit,
+  } = useForm<signTransactionFormData>({
+    defaultValues: {
+      destinationAddress: "0x08d2b0a37F869FF76BACB5Bab3278E26ab7067B7",
+    },
+  });
 
   // First, logout user if there is no current wallet set
   useEffect(() => {
@@ -54,34 +80,113 @@ export default function Home() {
     })();
   });
 
-  const signMessage = async (data: signingFormData) => {
+  useEffect(() => {
+    const initializeClients = async () => {
+      if (!wallet || !passkeyClient) return;
+
+      let smartAccount;
+      if (useViem) {
+        const viemAccount = await createAccount({
+          client: passkeyClient,
+          organizationId: wallet.subOrgId,
+          signWith: wallet.address,
+          ethereumAddress: wallet.address,
+        });
+
+        const viemClient = createWalletClient({
+          account: viemAccount as Account,
+          chain: sepolia,
+          transport: http(),
+        });
+
+        setViemClient(viemClient);
+
+        smartAccount = await connectViemClient(viemClient);
+      } else {
+        const ethersClient = new TurnkeySigner({
+          organizationId: wallet.subOrgId,
+          client: passkeyClient,
+          signWith: wallet.address,
+        });
+
+        const provider = new ethers.JsonRpcProvider(
+          `https://sepolia.infura.io/v3/${process.env.INFURA_KEY}`
+        );
+
+        const connectedSigner = ethersClient.connect(provider);
+
+        setEthersClient(connectedSigner);
+
+        smartAccount = await connectEthersClient(connectedSigner);
+      }
+
+      const smartAccountAddress = await smartAccount.getAccountAddress();
+      setSmartAccountAddress(smartAccountAddress);
+    };
+
+    initializeClients();
+  }, [wallet, passkeyClient, useViem]);
+
+  // Connect a TurnkeySigner to a Biconomy Smart Account Client, defaulting to Sepolia
+  const connectEthersClient = async (
+    turnkeyClient: TurnkeySigner
+  ): Promise<BiconomySmartAccountV2> => {
+    try {
+      const smartAccount = await createSmartAccountClient({
+        signer: turnkeyClient as LightSigner,
+        bundlerUrl: process.env.NEXT_PUBLIC_BICONOMY_BUNDLER_URL!, // <-- Read about this at https://docs.biconomy.io/dashboard#bundler-url
+        biconomyPaymasterApiKey:
+          process.env.NEXT_PUBLIC_BICONOMY_PAYMASTER_API_KEY!, // <-- Read about at https://docs.biconomy.io/dashboard/paymaster
+        rpcUrl: `https://sepolia.infura.io/v3/${process.env
+          .NEXT_PUBLIC_INFURA_KEY!}`, // <-- read about this at https://docs.biconomy.io/account/methods#createsmartaccountclient
+        chainId: Number(sepolia.id),
+      });
+
+      return smartAccount;
+    } catch (error: any) {
+      throw new Error(error);
+    }
+  };
+
+  // Connect a TurnkeySigner to a Biconomy Smart Account Client, defaulting to Sepolia
+  const connectViemClient = async (
+    turnkeyClient: WalletClient
+  ): Promise<BiconomySmartAccountV2> => {
+    try {
+      const smartAccount = await createSmartAccountClient({
+        signer: turnkeyClient,
+        bundlerUrl: process.env.NEXT_PUBLIC_BICONOMY_BUNDLER_URL!, // <-- Read about this at https://docs.biconomy.io/dashboard#bundler-url
+        biconomyPaymasterApiKey:
+          process.env.NEXT_PUBLIC_BICONOMY_PAYMASTER_API_KEY!, // <-- Read about at https://docs.biconomy.io/dashboard/paymaster
+        rpcUrl: `https://sepolia.infura.io/v3/${process.env
+          .NEXT_PUBLIC_INFURA_KEY!}`, // <-- read about this at https://docs.biconomy.io/account/methods#createsmartaccountclient
+        chainId: Number(sepolia.id),
+      });
+
+      return smartAccount;
+    } catch (error: any) {
+      throw new Error(error);
+    }
+  };
+
+  const signMessage = async (data: signMessageFormData) => {
     if (!wallet) {
       throw new Error("wallet not found");
     }
 
     if (useViem) {
+      if (!viemClient) throw new Error("viem client not initialized");
       await signMessageViem(data);
     } else {
+      if (!ethersClient) throw new Error("ethers client not initialized");
       await signMessageEthers(data);
     }
   };
 
-  const signMessageViem = async (data: signingFormData) => {
-    const viemAccount = await createAccount({
-      client: passkeyClient!,
-      organizationId: wallet!.subOrgId,
-      signWith: wallet!.address,
-      ethereumAddress: wallet!.address,
-    });
-
-    const viemClient = createWalletClient({
-      account: viemAccount as Account,
-      chain: sepolia,
-      transport: http(),
-    });
-
-    const signedMessage = await viemClient.signMessage({
+  const signMessageViem = async (data: signMessageFormData) => {
+    const signedMessage = await viemClient!.signMessage({
       message: data.messageToSign,
+      account: viemClient!.account!,
     });
 
     setSignedMessage({
@@ -90,14 +195,8 @@ export default function Home() {
     });
   };
 
-  const signMessageEthers = async (data: signingFormData) => {
-    const ethersSigner = new TurnkeySigner({
-      client: passkeyClient!,
-      organizationId: wallet!.subOrgId,
-      signWith: wallet!.address,
-    });
-
-    const signedMessage = await ethersSigner.signMessage(data.messageToSign);
+  const signMessageEthers = async (data: signMessageFormData) => {
+    const signedMessage = await ethersClient!.signMessage(data.messageToSign);
 
     setSignedMessage({
       message: data.messageToSign,
@@ -131,6 +230,7 @@ export default function Home() {
     });
 
     const response = res.data as TWalletDetails;
+
     setWallet(response);
 
     if (passkeyClient && passkeyClient.config) {
@@ -178,6 +278,94 @@ export default function Home() {
     }
   };
 
+  const signTransactionWithAAClient = async (
+    data: signTransactionFormData,
+    getSmartAccount: (
+      data: signTransactionFormData
+    ) => Promise<BiconomySmartAccountV2>,
+    options?: { nonceOptions?: { nonceKey: number } }
+  ) => {
+    const { destinationAddress, amount } = data;
+
+    const smartAccount = await getSmartAccount(data);
+    const transactionRequest = {
+      to: destinationAddress,
+      value: amount,
+      type: 2,
+    };
+
+    const userOpResponse = await smartAccount?.sendTransaction(
+      transactionRequest,
+      {
+        ...options,
+        paymasterServiceData: { mode: PaymasterMode.SPONSORED },
+      }
+    );
+
+    const { transactionHash } = await userOpResponse.waitForTxHash();
+
+    setSignedTransaction(
+      `https://jiffyscan.xyz/bundle/${transactionHash}?network=sepolia&pageNo=0&pageSize=10`
+    );
+  };
+
+  const signTransactionWithClient = async (
+    data: signTransactionFormData,
+    client: WalletClient | TurnkeySigner,
+    isViem: boolean
+  ) => {
+    const { destinationAddress, amount } = data;
+
+    const transactionRequest = {
+      to: destinationAddress as `0x${string}`,
+      value: BigInt(amount),
+      ...(isViem && {
+        account: (client as WalletClient).account!,
+        chain: sepolia,
+      }),
+    };
+
+    const result = await client.sendTransaction({
+      ...transactionRequest,
+      chain: transactionRequest.chain,
+      account: transactionRequest.account || null,
+    });
+    const txHash = isViem ? result : (result as any).hash;
+
+    setSignedTransaction(`https://sepolia.etherscan.io/tx/${txHash}`);
+  };
+
+  const signTransactionWithProvider = async (
+    data: signTransactionFormData,
+    client: WalletClient | TurnkeySigner,
+    isViem: boolean
+  ) => {
+    if (supportsAA()) {
+      await signTransactionWithAAClient(
+        data,
+        async () =>
+          isViem
+            ? connectViemClient(client as WalletClient)
+            : connectEthersClient(client as TurnkeySigner),
+        { nonceOptions: { nonceKey: Number(0) } }
+      );
+    } else {
+      await signTransactionWithClient(data, client, isViem);
+    }
+  };
+
+  const signTransaction = async (data: signTransactionFormData) => {
+    if (!wallet) {
+      throw new Error("wallet not found");
+    }
+
+    await signTransactionWithProvider(
+      data,
+      useViem ? viemClient! : ethersClient!,
+      useViem
+    );
+  };
+
   return (
     <main className={styles.main}>
       <a href="https://turnkey.com" target="_blank" rel="noopener noreferrer">
@@ -191,7 +379,7 @@ export default function Home() {
         />
       </a>
       <div>
-        {wallet !== null && (
+        {wallet && (
           <div className={styles.info}>
             Your sub-org ID: <br />
             <span className={styles.code}>{wallet.subOrgId}</span>
@@ -199,8 +387,14 @@ export default function Home() {
         )}
         {wallet && (
           <div className={styles.info}>
-            ETH address: <br />
+            ETH signer address: <br />
             <span className={styles.code}>{wallet.address}</span>
+          </div>
+        )}
+        {smartAccountAddress && (
+          <div className={styles.info}>
+            Smart account address: <br />
+            <span className={styles.code}>{smartAccountAddress}</span>
           </div>
         )}
         {signedMessage && (
@@ -366,6 +560,53 @@ export default function Home() {
           </form>
         </div>
       )}
+      {wallet && (
+        <div>
+          <h2>... and now let&apos;s sign a transaction!</h2>
+          <form
+            className={styles.form}
+            onSubmit={signTransactionFormSubmit(signTransaction)}
+          >
+            <input
+              className={styles.input}
+              {...signTransactionFormRegister("amount")}
+              placeholder="Amount (wei)"
+            />
+            <input
+              className={styles.input}
+              {...signTransactionFormRegister("destinationAddress")}
+            />
+            <input
+              className={styles.button}
+              type="submit"
+              value="Sign and Broadcast Transaction"
+            />
+          </form>
+          {signedTransaction && (
+            <div className={styles.info}>
+              <p>
+                🚀 Transaction broadcasted and confirmed!
+                <br />
+                <br />
+                <a
+                  href={signedTransaction}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                >
+                  {signedTransaction}
+                </a>
+              </p>
+            </div>
+          )}
+        </div>
+      )}
     </main>
   );
 }
+
+const supportsAA = () => {
+  return (
+    process.env.NEXT_PUBLIC_BICONOMY_BUNDLER_URL &&
+    process.env.NEXT_PUBLIC_BICONOMY_PAYMASTER_API_KEY
+  );
+};